Move the upload code into backups.go.
